requirements.txt
azure-functions
azure-storage
azure-storage-blob
flake8
requests